+++ core/src/main/scala/io/github/scalaquest/core/pipeline/resolver/Resolver.scala
@@ -74,31 +74,11 @@ object Resolver {
     s =>
       new SimpleResolver {
 
-        def isSubSet(d1: ItemDescription, d2: ItemDescription): Boolean = {
-          def decorators(d: ItemDescription): Set[String] = {
-            @tailrec
-            def go(d: ItemDescription, acc: Set[String]): Set[String] =
-              d match {
-                case DecoratedItem(decoration, item) => go(item, acc + decoration)
-                case _                               => acc
-              }
-            go(d, Set())
-          }
-          @tailrec
-          def base(d: ItemDescription): BaseItem =
-            d match {
-              case i: BaseItem         => i
-              case DecoratedItem(_, i) => base(i)
-            }
-
-          base(d1) == base(d2) && decorators(d1).subsetOf(decorators(d2))
-        }
-
         override def actions: PartialFunction[String, Action] = s.actions
 
         override def items: PartialFunction[ItemDescription, ItemRef] =
           d =>
-            s.game.itemsInScope.filter(i => isSubSet(d, i.description)).toList match {
+            s.game.itemsInScope.filter(i => d.isSubset(i.description)).toList match {
               case x :: Nil => x.itemRef
             }
       }
